Improve the indentation script.
Now, something like:

\ifthenelse{...}{
	Yes.
}{
	No.
}

is properly indented (as above). The third line is indented according to the
position in the line which causes the least indentation (after as many braces
as possible were closed).

We also try to indent \if/\else/\fi blocks:
\ifnum1<0
	bar
\else
	foo
\fi

Finally, the indentation is much faster (by using a cache) than the default
indentation script shipped with vim.