Change behaviour of cite-completion
Now, if you hit <F9> on a line which does not contain a key of an entry, it will
complete with the previous entry, and not the following one. In
particular, if the cursor is on the second line of the entry, which is
typically the title, the previous version completed with the following
entry.

The new behaviour seems to be more intuitive.