BugFix: Folding: Section Title Parsing
Previously, if a section began anywhere other than the first column of
the line, then the title of that section would usually be omitted from
the fold summary.

e.g.

```
\documentclass{article}
\begin{document}
\section{This title was always read}
  Lorem ipsum
  \section{This title was not read}
  Lorem ipsum
  \section*{This title was also not read}
  Lorem ipsum
\end{document}
```

This error occurred because of a mismatch in how whitespace was handled
between the fold locator and the section title parser.

Now the section title is always detected.